+++ R/reexports.R
@@ -1,9 +1,3 @@
-#' @title Re-export of `as.data.table`
-#'
-#' See [data.table::as.data.table].
-#'
-#' @name as.data.table
-#' @rdname as.data.table
-#' @keywords internal
+#' @importFrom data.table as.data.table
 #' @export
-NULL
+data.table::as.data.table